+++ test/hub/test_hub.dart
@@ -1,0 +1,15 @@
+import 'package:datahub/datahub.dart';
+import 'package:datahub/rest_client.dart';
+
+import '../transfer_object/contact.dart';
+
+part 'test_hub.g.dart';
+
+@Hub()
+abstract class TestHub {
+  @HubResource('/contact/{id}')
+  Resource<Contact> get contact;
+
+  @HubResource('/self')
+  MutableResource<Contact> get self;
+}